+++ lib/carRight.js
@@ -1,3 +1,7 @@
+var canvas = document.getElementById("frogger");
+var ctx = canvas.getContext("2d");
+var rightCarImg = document.getElementById("right-car-img");
+
 function CarRight(x, y, vx, width){
   this.x = x;
   this.y = y;
@@ -8,7 +12,7 @@ function CarRight(x, y, vx, width){
 
 
 
-CarRight.prototype.draw = function (ctx, rightCarImg) {
+CarRight.prototype.draw = function () {
   ctx.fillStyle = 'transparent';
   ctx.drawImage(rightCarImg, this.x, this.y, this.width, this.height);
   return this;
@@ -22,6 +26,7 @@ CarRight.prototype.move = function () {
   return this;
   }
 };
+
 
 
 module.exports = CarRight;
