+++ lib/frog.js
@@ -1,3 +1,8 @@
+var canvas = document.getElementById("frogger");
+var ctx = canvas.getContext("2d");
+var frogImg = document.getElementById("frog-img");
+
+
 var Frog = function(x, y, height, width) {
   this.x = x;
   this.y = y;
@@ -5,7 +10,7 @@ var Frog = function(x, y, height, width) {
   this.width = width;
 };
 
-Frog.prototype.drawFrog = function(ctx, frogImg) {
+Frog.prototype.drawFrog = function() {
   ctx.beginPath();
   ctx.drawImage(frogImg, this.x, this.y, this.width, this.height);
   ctx.fillStyle = "transparent";
@@ -34,5 +39,10 @@ Frog.prototype.frogDown = function(){
   this.y += 50;
  }
 };
+
+Frog.prototype.resetFrog = function(){
+  this.x = (600-50)/2;
+  this.y = 700 -100;
+}
 
 module.exports = Frog;
