+++ lib/index.js
@@ -1,24 +1,18 @@
 var World = require('./world.js');
+var Collision = require('./collision.js');
+var Frog = require('./frog.js');
 var canvas = document.getElementById("frogger");
 var ctx = canvas.getContext("2d");
 var width = canvas.width;
 var height = canvas.height;
 var world = new World(canvas, ctx, width, height);
-
-canvas.focus();
-
+var collision = new Collision();
 
 requestAnimationFrame(function gameLoop() {
   ctx.clearRect(0, 0, width, height);
-  world.objectLoops();
+  collision.logLoops();
   world.drawings();
-  world.leftCarCollision();
-  world.rightCarCollision();
-  world.frogCollision();
-  world.logCollusion();
-  world.turtleCollusion();
-  world.lilyPadCollusion();
-  world.waterDeath();
+  world.collisionCheck();
   world.checkLives();
   world.startTheScreen();
   world.winScreen();
