+++ lib/lily-pad.js
@@ -1,3 +1,6 @@
+var canvas = document.getElementById("frogger");
+var ctx = canvas.getContext("2d");
+
 function LilyPad(x, y){
   this.x = x;
   this.y = y;
@@ -5,7 +8,7 @@ function LilyPad(x, y){
   this.width = 20;
 }
 
-LilyPad.prototype.draw = function(ctx){
+LilyPad.prototype.draw = function(){
   ctx.fillStyle = "transparent";
   ctx.fillRect(this.x, this.y, this.width, this.height);
 };
