+++ lib/logs.js
@@ -1,3 +1,8 @@
+var canvas = document.getElementById("frogger");
+var ctx = canvas.getContext("2d");
+var thinLogImg = document.getElementById("thin-log-img");
+
+
 function Log(x, y, vx, width){
   this.x = x;
   this.y = y;
@@ -6,15 +11,15 @@ function Log(x, y, vx, width){
   this.width = width;
 }
 
-Log.prototype.draw = function (ctx, thinLogImg) {
+Log.prototype.draw = function () {
   ctx.drawImage(thinLogImg, this.x, this.y, this.width, this.height);
   ctx.fillStyle = "transparent";
   return this;
 };
 
-Log.prototype.move = function(width) {
+Log.prototype.move = function() {
   this.x += this.vx;
-  if(this.x > width + 50){
+  if(this.x > 600 + 50){
     this.x = -200;
   return this;
   }
