+++ lib/turtles.js
@@ -1,3 +1,7 @@
+var canvas = document.getElementById("frogger");
+var ctx = canvas.getContext("2d");
+var turtleImg = document.getElementById("turtle-img");
+
 function Turtle(x, y, vx, width){
   this.x = x;
   this.y = y;
@@ -7,7 +11,7 @@ function Turtle(x, y, vx, width){
 }
 
 
-Turtle.prototype.draw = function (ctx, turtleImg) {
+Turtle.prototype.draw = function () {
   ctx.drawImage(turtleImg, this.x, this.y, this.width, this.height);
   ctx.fillStyle = 'transparent';
   return this;
