+++ lib/world.js
@@ -9,17 +9,15 @@ var GameOver = require('./gameover.js');
 var StartScreen = require('./startscreen.js');
 var WinScreen = require('./winscreen.js');
 var Music = require('./music.js');
+var Collision = require('./collision.js');
 
 var canvas = document.getElementById("frogger");
 var ctx = canvas.getContext("2d");
 var width = canvas.width;
 var height = canvas.height;
 
-var rightCarImg = document.getElementById("right-car-img");
-var leftCarImg = document.getElementById("left-car-img");
+
 var frogImg = document.getElementById("frog-img");
-var turtleImg = document.getElementById("turtle-img");
-var thinLogImg = document.getElementById("thin-log-img");
 var deadFrogImg = document.getElementById("dead-frog-img");
 var startScreenImg = document.getElementById("start-screen-img");
 var winImg = document.getElementById("win-img");
@@ -38,12 +36,10 @@ var World = function(){
 };
 
 var frog = new Frog((width-50)/2, height -100, 45, 50);
+var collision = new Collision();
+var carLeft = new CarLeft();
+var carRight = new CarRight();
 
-var frogs = [];
-frogs.push(new Frog(-5000, 0, 45, 50));
-frogs.push(new Frog(-5000, 0, 45, 50));
-frogs.push(new Frog(-5000, 0, 45, 50));
-frogs.push(new Frog(-5000, 0, 45, 50));
 
 var frogA = new Frog(550, 650, 45, 50);
 var frogB = new Frog(500, 650, 45, 50);
@@ -57,20 +53,28 @@ var music = new Music();
 
 var updateScore = 0;
 var counter = 0;
-var lives = 3;
+var lives = collision.lives;
 var start = false;
 var collide = false;
 var death = false;
 
+World.prototype.collisionCheck = function () {
+  collision.logCollusion(frog);
+  collision.turtleCollusion(frog);
+  collision.leftCarCollision(frog, lives);
+  collision.rightCarCollision(frog, lives);
+  collision.frogCollision(frog, lives);
+  collision.lilyPadCollusion(frog);
+  collision.waterDeath(frog, lives);
+  collide = collision.collide;
+};
+
 World.prototype.drawings = function() {
   frog.drawFrog(ctx, frogImg);
-  frogs.forEach(function(frogs, i){
-    frogs.drawFrog(ctx, frogImg);
-    frogA.drawFrog(ctx, frogImg);
-    frogB.drawFrog(ctx, frogImg);
-    frogC.drawFrog(ctx, frogImg);
-    score.draw(ctx, updateScore);
-  });
+  score.draw(ctx, updateScore);
+  frogA.drawFrog(ctx, frogImg);
+  frogB.drawFrog(ctx, frogImg);
+  frogC.drawFrog(ctx, frogImg);
 };
 
 document.addEventListener('keydown', function(e) {
@@ -99,183 +103,18 @@ document.addEventListener('keydown', function(e) {
   }
 });
 
-World.prototype.objectLoops = function() {
-  leftCar.forEach(function(car){
-    car.draw(ctx, leftCarImg).move(width);
-  });
-  rightCar.forEach(function(car){
-    car.draw(ctx, rightCarImg).move(width);
-  });
-  allLogs.forEach(function(log){
-    log.draw(ctx, thinLogImg).move(width);
-  });
-  allTurtles.forEach(function(turtle){
-    turtle.draw(ctx, turtleImg).move();
-  });
-  allLilyPads.forEach(function(lilypad){
-    lilypad.draw(ctx);
-  });
-};
-
-var leftCar = [];
-var rightCar = [];
-leftCar.push(new CarLeft(50, 550, 1, 50));
-leftCar.push(new CarLeft(250, 550, 1, 50));
-leftCar.push(new CarLeft(450, 550, 1, 50));
-rightCar.push(new CarRight(600, 500, -1, 50));
-rightCar.push(new CarRight(400, 500, -1, 50));
-rightCar.push(new CarRight(50, 500, -1, 50));
-leftCar.push(new CarLeft(0, 450, 1, 50));
-leftCar.push(new CarLeft(100, 450, 1, 50));
-leftCar.push(new CarLeft(400, 450, 1, 50));
-rightCar.push(new CarRight(400, 400, -1, 50));
-rightCar.push(new CarRight(150, 400, -1, 50));
-leftCar.push(new CarLeft(300, 350, 1, 50));
-leftCar.push(new CarLeft(100, 350, 1, 50));
-
-var allLogs = [];
-allLogs.push(new Log(-200, 200, 1, 100));
-allLogs.push(new Log(50, 200, 1, 100));
-allLogs.push(new Log(450, 200, 1, 100));
-allLogs.push(new Log(350, 150, 2, 200));
-allLogs.push(new Log(0, 150, 2, 200));
-allLogs.push(new Log(-200, 50, 1.5, 150));
-allLogs.push(new Log(100, 50, 1.5, 150));
-allLogs.push(new Log(450, 50, 1.5, 150));
-
-var allTurtles = [];
-allTurtles.push(new Turtle(-100, 251, -1, 150));
-allTurtles.push(new Turtle(250, 251, -1, 150));
-allTurtles.push(new Turtle(600, 251, -1, 150));
-allTurtles.push(new Turtle(-150, 100, -1.5, 150));
-allTurtles.push(new Turtle(50, 100, -1.5, 150));
-allTurtles.push(new Turtle(350, 100, -1.5, 150));
-allTurtles.push(new Turtle(650, 100, -1.5, 150));
-
-var allLilyPads = [];
-allLilyPads.push(new LilyPad(90, 0));
-allLilyPads.push(new LilyPad(190, 0));
-allLilyPads.push(new LilyPad(290, 0));
-allLilyPads.push(new LilyPad(390, 0));
-allLilyPads.push(new LilyPad(490, 0));
-
-function incrementSpeeds(){
-  leftCar.forEach(function(car){
-    car.vx = car.vx + 0.5;
-  });
-  rightCar.forEach(function(car){
-    car.vx = car.vx - 0.5;
-  });
-}
-
-function resetSpeeds(){
-  leftCar.forEach(function(car){
-    car.vx = 1;
-  });
-  rightCar.forEach(function(car){
-    car.vx = -1;
-  });
-}
-
-function frogAndCarCollision(item){
-  if(frog.x < item.x + item.width &&
-   frog.x + frog.width > item.x &&
-   frog.y < item.y + item.height &&
-   frog.height + frog.y > item.y) {
-    resetFrog();
-    lives--;
-    }
-}
-
-World.prototype.frogCollision = function(){
-  frogs.forEach(function(item, i){
-  frogAndCarCollision(item);
-  });
-};
-
-World.prototype.leftCarCollision = function(){
-  leftCar.forEach(function(item, i){
-  frogAndCarCollision(item);
-  });
-};
-
-World.prototype.rightCarCollision = function() {
-  rightCar.forEach(function(item, i){
-frogAndCarCollision(item);
-  });
-};
-
-function logAndTurtleCollision(item){
-  if(frog.x < item.x + item.width - 25 &&
-  frog.x + frog.width > item.x + 25 &&
-  frog.y < item.y + item.height &&
-  frog.height + frog.y > item.y) {
-    (frog.x = frog.x + item.vx);
-    (frog.y = item.y + 1);
-    collide = true;
-  }
-}
-
-World.prototype.logCollusion = function() {
-    allLogs.forEach(function(item, i){
-      logAndTurtleCollision(item);
-    });
-  };
-
-World.prototype.turtleCollusion = function() {
-  allTurtles.forEach(function(item, i){
-    logAndTurtleCollision(item);
-  });
-};
-
-World.prototype.lilyPadCollusion = function() {
-    allLilyPads.forEach(function(lilypad, i){
-    if (frog.x < lilypad.x + lilypad.width &&
-      frog.x + frog.width > lilypad.x &&
-      frog.y < lilypad.y + lilypad.height &&
-      frog.height + frog.y > lilypad.y) {
-        (frog.x = lilypad.x -15);
-        (frog.y = lilypad.y + 1);
-        collide = true;
-        resetFrog();
-        counter ++;
-        if(counter === 1) {
-          frogs[0].x = lilypad.x-15;
-          incrementSpeeds();
-        }
-        else if(counter === 2) {
-          frogs[1].x = lilypad.x-15;
-          incrementSpeeds();
-        }
-        else if(counter === 3) {
-          frogs[2].x = lilypad.x-15;
-          incrementSpeeds();
-        }
-        else if(counter === 4) {
-          frogs[3].x = lilypad.x-15;
-          incrementSpeeds();
-        }
-      return updateScore += 90;
-    }
-  });
-};
-
-function resetFrog(){
-  frog.x = (width-50)/2;
-  frog.y = height -100;
-}
-
-World.prototype.waterDeath = function () {
-  if(frog.y < 300) {
-  if(collide === false){
-    resetFrog();
-    lives--;
-    }
-  else if(collide === true){
-    collide = false;
-    }
-  }
-};
+// World.prototype.waterDeath = function () {
+//   if(frog.y < 300) {
+//   if(collide === false){
+//       console.log('this is our collision', self.collide)
+//     frog.resetFrog();
+//     lives--;
+//     }
+//   else if(collide === true){
+//     collide = false;
+//     }
+//   }
+// };
 
 function spaceBarReload(){
   document.addEventListener('keydown', function(e) {
@@ -291,12 +130,10 @@ function resetDefaults(){
   frogA.width = 50;
   frogB.width = 50;
   frogC.width = 50;
-  frogs.forEach(function(frogs, i){
-    frogs.x = -5000;
-  });
+  collision.resetPlaceholders();
   updateScore = 0;
-  resetFrog();
-  resetSpeeds();
+  frog.resetFrog();
+  collision.resetSpeeds();
   start = true;
   death = false;
   counter = 0;
